deployment/requirements_dev.txt
fastapi
uvicorn
pydantic==1.10
typing
python-dotenv
requests
sqlalchemy
fastapi-crudrouter
typing_extensions<4.6.0
python-multipart